+++ CHANGELOG.md
@@ -7,7 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
-- Added Stats gathering in servers
+- Added Stats ([!128](https://storage.bsc.es/gitlab/hpc/gekkofs/-/merge_requests/128)) gathering in servers
+- GKFS_CHUNK_STATS enables chunk usage output
+- Stats output can be enabled with --output-stats <filename>
 ### New
 
 - Added new experimental metadata backend: